requirements.txt
# Requirements for Market Data Extraction Project
# Install with: pip install -r requirements.txt
# Updated for Phase 5: Analysis and Visualization

# ==================== CORE DATA MANIPULATION ====================
pandas>=2.0.0
numpy>=1.24.0

# ==================== FINANCIAL DATA ====================
yfinance>=0.2.28

# ==================== HTTP REQUESTS FOR APIs ====================
requests>=2.31.0

# ==================== ENVIRONMENT VARIABLES ====================
python-dotenv>=1.0.0

# ==================== DATE UTILITIES ====================
python-dateutil>=2.8.2

# ==================== VISUALIZATION (PHASE 5) ====================
# Required for plots_report() and all visualizations
matplotlib>=3.7.0
seaborn>=0.12.0

# ==================== JUPYTER NOTEBOOKS (PHASE 5) ====================
# Required for portfolio_analysis.ipynb
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# ==================== OPTIONAL BUT RECOMMENDED ====================
# Progress bars for long operations
tqdm>=4.65.0

# Data validation
pydantic>=2.0.0

# Enhanced plotting capabilities (optional)
plotly>=5.14.0

# For better notebook rendering (optional)
nbformat>=5.9.0
